Fix RKNN INT8 coordinate restoration
