+++ ultralytics/nn/backends/rknn.py
@@ -60,17 +60,16 @@ class RKNNBackend(BaseBackend):
         """Run inference on the Rockchip NPU.
 
         Args:
-            im (torch.Tensor): Input image tensor in BCHW format, normalized to [0, 1].
+            im (torch.Tensor): Input image tensor in BHWC format, normalized to [0, 1].
 
         Returns:
             (list): Model predictions as a list of output arrays.
         """
-        h, w = im.shape[2:4]
+        h, w = im.shape[1:3]
         im = (im.cpu().numpy() * 255).astype("uint8")
         im = im if isinstance(im, (list, tuple)) else [im]
         y = self.model.inference(inputs=im)
-        # INT8 exports normalize box (and pose keypoint) coordinates to [0, 1] so a single per-tensor scale does not
-        # collapse the class scores (see _NormalizeCoords); undo it here. Floating-point builds are already in pixels.
+        # INT8 exports use input-relative coordinates so a single per-tensor scale preserves class scores.
         if (
             self.metadata.get("args", {}).get("quantize") == 8
             and self.task in {"detect", "segment", "pose", "obb"}
